+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+- Error when SKU has no sellers.
+- Duplicated values for SKU specifications.
+
 ## [1.6.3] - 2024-04-01
 
 ### Fixed